fix: GUI object store timeouts in kubernetes.
HttpFrontend in kubernetes was unable to resolve url received from the datalayer. Now HttpFrontend for kubernetes will connect to management service with a Host header set as the hostname dervived from the management service httpurl obtained from datalayer and will connect to a specialized http url dervied from the first part of the hostname prefixed with http.