Improving deployment documentation
main README now refers to the Helm package README, removed copy and pasted description to avoid two locations to maintain the description.
helm package README changed to accomodate for the missing main Makefile and the lack of a global make description.
Added descriptions how to change the private container registry.