Helm: Istio ingress for Kubernetes deployments
Issue #23 discussed avoiding CORS checks by using our Nginx that hosts the GUI pages as a reverse proxy for HttpFrontend and the management workflow
Issue #44 notes that Nginx can not reliably determine the management workflow upstream and may result in an error state if the management service does not come up in time for nginx to resolve it as upstream

Using Istio for all ingress routing does provide a solution
The current configuration reuses the knative-serving/knative-ingress-gateway and presumes that istio-system/istio-ingressgateway is used for incoming traffic. The ingress does not install a certificate to secure traffic, but it configures the virtualservice to forward traffic to /, /storage and /management to their respective backends GUI (nx service), HttpFrontend (NEW store service) and the management workflow (wf-<rel>-management). The dependency on the knative-ingress-gateway as well as the use of FQDNs in forwarding knative service traffic is currently necessary, all other trials resulted in 503 service unavailable of 404 not found when invoking the management service
Next step would be to test the Kubernetes Ingress resource as a generic solution